# Conflicts:
#	src/MarginTrading.Backend.Core/Settings/MarginTradingSettings.cs
#	src/MarginTrading.Backend.Services/AssetPairs/ScheduleSettingsCacheService.cs
#	src/MarginTrading.Backend.Services/EventsConsumers/TradesConsumer.cs
#	src/MarginTrading.Backend.Services/Infrastructure/ConfigurationValidator.cs
#	src/MarginTrading.Backend.Services/Infrastructure/IConfigurationValidator.cs
#	src/MarginTrading.Backend.Services/MarginTrading.Backend.Services.csproj
#	src/MarginTrading.Backend.Services/Modules/CqrsModule.cs
#	src/MarginTrading.Backend.Services/Modules/ServicesModule.cs
#	src/MarginTrading.Backend/Middleware/GlobalErrorHandlerMiddleware.cs
#	src/MarginTrading.Backend/Middleware/ValidationExceptionHandler.cs
#	src/MarginTrading.Common/Helpers/StreamHelpers.cs
#	src/MarginTrading.Common/MarginTrading.Common.csproj
#	src/MarginTrading.Common/Middleware/RequestsLoggingMiddleware.cs
#	tests/MarginTradingTests/AssetDayOffTests.cs
#	tests/MarginTradingTests/StreamPartialReaderTest.cs

+++ src/MarginTrading.Common/Middleware/RequestsLoggingMiddleware.cs
@@ -55,7 +55,7 @@ namespace MarginTrading.Common.Middleware
 
                     using (originalRequestBody)
                     {
-                        var bytes = await originalRequestBody.ReadBytes(_settings.MaxPartSize);
+                        var bytes = await originalRequestBody.ReadBytes((uint)_settings.MaxPartSize);
                         var body = bytes == null ? null : System.Text.Encoding.UTF8.GetString(bytes);
                         var headers = context.Request.Headers.Where(h => !_personalDataHeaders.Contains(h.Key)).ToJson();
                         var userId = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
